UPDATE:
Added a number of new drivers to solve Hubbard and BHZ model with
different unit cells (for the AFM regime). Also new drivers to
solve Graphene and Haldane model.

- hm_2d: solve the HM in 2D (used as a benchmark)
- hm_2d_afm2: solve the HM in 2D with two atoms in the unit cell
- hm_2d_afm4: solve the HM in 2D with four atoms in the unit cell

- bhz_2d_afm2: solve the BHZ in 2D with two atoms in the unit cell
- bhz_2d_afm4: (updated) solve the BHZ in 2D with two atoms in the unit cell
- bhz_2d_edge_afm2: solve the BHZ in a 2D-Stripe with two atoms in the unit cell
- bhz_3d_afm8: solve the BHZ in 3D with 8 atoms in the unit cell

- graphene: solve the Graphene model in 2D
- graphene: solve the Graphene model in 2D with two atoms in the unit cell (already there huh!)
- haldane: solve the Haldane model

tests are in ~/scratch/testing/test_tb and all give perfect results: complete match between
different equivalent codes.

*update*:
updated the driver bhz_2d_edge_afm2 to be compliant with the DMFT driver.
The two give the same results. For this choice of the Hk,G_k structure the
tridiagonal inversion algorithm CAN NOT be used.